services/projects/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.5.0
httpx==0.27.2
pydantic==2.6.4
sqlalchemy==2.0.29
psycopg[binary]==3.1.19
